Move DANCE_STYLE_CHOICES outside of Heat model
