+++ comps/views/heatlist_error_crud.py
@@ -1,7 +1,7 @@
 from django.core.paginator import Paginator
 from django.shortcuts import redirect, render, get_object_or_404
 from comps.models.comp import Comp
-from comps.models.heat import Heat
+from comps.models.heat import Heat, UNKNOWN
 from comps.models.heatlist_error import Heatlist_Error
 
 
@@ -19,7 +19,7 @@ def check_heatlist_error(request, error_id):
         if heat.base_value > 0:
             heatlist_error.delete()
     elif heatlist_error.error == Heatlist_Error.UNKNOWN_STYLE:
-        if heat.style != Heat.UNKNOWN:
+        if heat.style != UNKNOWN:
             heatlist_error.delete()
     return redirect('comps:heat', heat.id)
