+++ rankings/views/calc_rankings.py
@@ -1,10 +1,10 @@
 from django.core.cache import cache
 from django.core.paginator import Paginator
 from django.shortcuts import render, redirect, get_object_or_404
-from rankings.models import Couple
+from rankings.models.couple import Couple
 from rankings.tasks import calc_couple_ratings
 from comps.models.comp import Comp
-from comps.models.heat import Heat
+from comps.models.heat import Heat, DANCE_STYLE_CHOICES
 from comps.models.heat_entry import Heat_Entry
 from operator import itemgetter
 
@@ -18,7 +18,7 @@ def calc_rankings(request):
     for c in couple_types:
         couple_type_labels.append(c[1])
 
-    styles = Heat.DANCE_STYLE_CHOICES
+    styles = DANCE_STYLE_CHOICES
     style_choices = list()
     for s in styles:
         style_choices.append(s[0])
@@ -51,7 +51,7 @@ def calc_rankings(request):
         heat_couple_type = Couple.COUPLE_TYPE_CHOICES[index][0]
         style = request.POST.get("style")
         index = style_labels.index(style)
-        heat_style = Heat.DANCE_STYLE_CHOICES[index][0]
+        heat_style = DANCE_STYLE_CHOICES[index][0]
         current_url = request.path
         url_string = current_url +"?type=" + heat_couple_type + "&style=" + heat_style
         last_name = request.POST.get("last_name")
@@ -61,6 +61,7 @@ def calc_rankings(request):
 
     # try to read the ranking data from the cache
     couple_stats = cache.get(cache_key)
+    
 
     if couple_stats is None:
         # cache miss - calculate rankings and store in cache
